requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.6
Django==3.2.19
django-resized==1.0.2
django-threaded-email==0.0.1
filelock==3.12.0
idna==3.4
packaging==23.1
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
pypng==0.20220715.0
pyproject_api==1.5.1
pytz==2023.3
qrcode==7.4.2
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
tox==4.5.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.21.0
